+++ src/modules/Elsa.Workflows.Api/Endpoints/RuntimeAdmin/Models.cs
@@ -1,4 +1,6 @@
-namespace Elsa.Workflows.Runtime.Admin;
+using Elsa.Workflows.Runtime;
+
+namespace Elsa.Workflows.Api.Endpoints.RuntimeAdmin;
 
 /// <summary>Body for the pause endpoint. The reason is a free-form human-readable string captured on the audit event.</summary>
 public class PauseRequest
