--- src/modules/Elsa.Workflows.Runtime.Admin/Elsa.Workflows.Runtime.Admin.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <Description>
-            Provides the administrative HTTP API for the workflow runtime's graceful-shutdown surface (pause, resume, status, force).
-        </Description>
-        <PackageTags>elsa module workflows runtime admin graceful-shutdown</PackageTags>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="CShells" />
-        <PackageReference Include="CShells.FastEndpoints.Abstractions" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\..\common\Elsa.Api.Common\Elsa.Api.Common.csproj" />
-        <ProjectReference Include="..\Elsa.Workflows.Runtime\Elsa.Workflows.Runtime.csproj" />
-    </ItemGroup>
-
-</Project>
